:truck: Move ContextSettings to config
