+++ data_safe_haven/config/__init__.py
@@ -1,9 +1,11 @@
+from .context_settings import ContextSettings
 from .dsh_pulumi_config import DSHPulumiConfig
 from .dsh_pulumi_project import DSHPulumiProject
 from .shm_config import SHMConfig
 from .sre_config import SREConfig
 
 __all__ = [
+    "ContextSettings",
     "DSHPulumiConfig",
     "DSHPulumiProject",
     "SHMConfig",
