+++ data_safe_haven/config/context_settings.py
@@ -11,6 +11,7 @@ from typing import ClassVar
 
 from pydantic import Field, model_validator
 
+from data_safe_haven.context import Context
 from data_safe_haven.directories import config_dir
 from data_safe_haven.exceptions import (
     DataSafeHavenConfigError,
@@ -18,8 +19,6 @@ from data_safe_haven.exceptions import (
 )
 from data_safe_haven.logging import get_logger
 from data_safe_haven.serialisers import YAMLSerialisableModel
-
-from .context import Context
 
 
 class ContextSettings(YAMLSerialisableModel):
