+++ data_safe_haven/context/__init__.py
@@ -1,7 +1,5 @@
 from .context import Context
-from .context_settings import ContextSettings
 
 __all__ = [
     "Context",
-    "ContextSettings",
 ]
